requirements.txt
streamlit==0.86.0
ibm_cloud_sdk_core==3.11.3
ibm_watson==5.3.0
pandas==1.2.4
python-dotenv==0.19.1
